feat : 브리더 신고 구현
